Binary Interactions

Below is a list of the currently available binary interactions:

Collision of hard spheres

Photon pair production from electron positron annihilation

Electron positron pair production from photon pair annihilation

Differential and total cross section functions:

BinaryInteractionSpectra.sigma_SphSphSphSph — Function

sigma_SphSphSphSph(sSmol,sBig)

returns the total cross section for the binary interaction of hard spheres with normalised masses m1,m2,m3,m4

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2

BinaryInteractionSpectra.dsigmadt_SphSphSphSph — Function

dsigmadt_SphSphSphSph(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for the binary interaction of hard spheres with normalised masses m1,m2,m3,m4

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2

tSmol::Float32 : t - tBig

tBig::Float32 : (m3-m1)^2

uSmol::Float32 : u - uBig

uBig::Float32 : (m2-m3)^2

BinaryInteractionSpectra.sigma_ElePosPhoPho — Function

sigma_ElePosPhoPho(sSmol,sBig)

returns the total cross section for electron positron annihilation to two photons. Berestetskii 1982 (88.6)

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 4 (normalised units) -> s = sSmol + 4

BinaryInteractionSpectra.dsigmadt_ElePosPhoPho — Function

dsigmadt_ElePosPhoPho(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for electron positron annihilation to two photons. Berestetskii 1982 (88.4)

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 4 (normalised units) -> s = sSmol + 4

tSmol::Float32 : t - tBig

tBig::Float32 : (m3-m1)^2 = 1 (normalised units) -> t = tSmol + 1

uSmol::Float32 : u - uBig

uBig::Float32 : (m2-m3)^2 = 1 (normalised units) -> u = uSmol + 1

BinaryInteractionSpectra.sigma_PhoPhoElePos — Function

sigma_PhoPhoElePos(sSmol,sBig)

returns the total cross section for photon-photon annihilation to electron-positron pair.

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 0 (normalised units) -> s = sSmol

BinaryInteractionSpectra.dsigmadt_PhoPhoElePos — Function

dsigmadt_PhoPhoElePos(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for photon-photon annihilation to electron-positron pair.

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 0 (normalised units) -> s = sSmol

tSmol::Float32 : t - tBig

tBig::Float32 : (m3-m1)^2 = 1 (normalised units) -> t = tSmol + 1

uSmol::Float32 : u - uBig

uBig::Float32 : (m2-m3)^2 = 1 (normalised units) -> u = uSmol + 1