Cross Sections

Below is a list of the currently available binary interaction cross sections that have been implemented:

Collision of hard spheres SphSphSphSph

Photon pair production from electron positron annihilation ElePosPhoPho

Electron positron pair production from photon pair annihilation PhoPhoElePos

Differential and total cross section functions:

All cross sections are to be defined in terms of the Mandelstram variables $s=(p_1^\mu+p_2^\mu)^2$, $t=(p_1^\mu-p_3^\mu)^2$ and $u=(p_2^\mu-p^3^\mu)^2. To maintain accuracy of cross sections and avoid DivZero issues when momenta is small compared to the mass of the particles (at Float32 precision), each Mandelstram variable is split into two components e.g. $s=sSmol+sBig$ where $sBig = (m_1+m_2)^2$. The latter part typically cancels with terms in the cross sections, leading to better accuracy.

BinaryInteractionSpectra.sigma_SphSphSphSph — Function

sigma_SphSphSphSph(sSmol,sBig)

returns the total cross section for the binary interaction of hard spheres with normalised masses m1,m2,m3,m4

\[σ = \frac{1}{2}\]

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2

BinaryInteractionSpectra.dsigmadt_SphSphSphSph — Function

dsigmadt_SphSphSphSph(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for the binary interaction of hard spheres with normalised masses m1,m2,m3,m4

\[\frac{dσ}{dt} = \frac{1}{s-4μ_{ ext{Sph}}^2}\]

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2

tSmol::Float32 : t - tBig

tBig::Float32 : (m3-m1)^2

uSmol::Float32 : u - uBig

uBig::Float32 : (m2-m3)^2

BinaryInteractionSpectra.sigma_ElePosPhoPho — Function

sigma_ElePosPhoPho(sSmol,sBig)

returns the total cross section for electron positron annihilation to two photons. Berestetskii 1982 (88.6)

\[σ = \frac{1}{4s^2(s-4)}\left((s^2+4s-8)\log\left(\frac{\sqrt{s}+\sqrt{s-4}}{\sqrt{s}-\sqrt{s-4}}\right)-(s+4)\sqrt{s(s-4)}\right)\]

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 4 (normalised units) -> s = sSmol + 4

BinaryInteractionSpectra.dsigmadt_ElePosPhoPho — Function

dsigmadt_ElePosPhoPho(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for electron positron annihilation to two photons. Berestetskii 1982 (88.4)

\[\frac{dσ}{dt} = -\frac{1}{s(s-4)}\left(\left(\frac{1}{t-1}+\frac{1}{u-1}\right)^2+\left(\frac{1}{t-1}+\frac{1}{u-1}\right)-\frac{1}{4}\left(\frac{t-1}{u-1}+\frac{u-1}{t-1}\right)\right)\]

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 4 (normalised units) -> s = sSmol + 4

tSmol::Float32 : t - tBig

tBig::Float32 : (m3-m1)^2 = 1 (normalised units) -> t = tSmol + 1

uSmol::Float32 : u - uBig

uBig::Float32 : (m2-m3)^2 = 1 (normalised units) -> u = uSmol + 1

BinaryInteractionSpectra.sigma_PhoPhoElePos — Function

sigma_PhoPhoElePos(sSmol,sBig)

returns the total cross section for photon-photon annihilation to electron-positron pair.

\[sigma = \frac{1}{2s^3}\left((s^2+4s-8)\log\left(\frac{\sqrt(s)+\sqrt(s-4)}{\sqrt(s)-\sqrt(s-4)}\right)-(s+4)\sqrt{s(s-4)}\right)\]

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 0 (normalised units) -> s = sSmol

BinaryInteractionSpectra.dsigmadt_PhoPhoElePos — Function

dsigmadt_PhoPhoElePos(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for photon-photon annihilation to electron-positron pair. (Inverse proceess of electron positron annihilation to two photons)

\frac{dσ}{dt} = -\frac{1}{s^2}\left(\left(\frac{1}{t-1}+\frac{1}{u-1}\right)^2+\left(\frac{1}{t-1}+\frac{1}{u-1}\right)-\frac{1}{4}\left(\frac{t-1}{u-1}+\frac{u-1}{t-1}\right)\right)

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2 = 0 (normalised units) -> s = sSmol

tSmol::Float32 : t - tBig

tBig::Float32 : (m3-m1)^2 = 1 (normalised units) -> t = tSmol + 1

uSmol::Float32 : u - uBig

uBig::Float32 : (m2-m3)^2 = 1 (normalised units) -> u = uSmol + 1