Internal Functions

BoltzmannCollisionIntegral.InvarientFlux — Method

InvarientFlux(s,mu12,mu22)

returns the value of the invarient flux with 's' mandelstram variable and masses 'mass1' and 'mass2'

BoltzmannCollisionIntegral.InvarientFlux2 — Method

InvarientFlux2(s,mass12,mass22)

returns the value of the squared invarient flux with 's' mandelstram variable and masses 'mass1' and 'mass2'

BoltzmannCollisionIntegral.InvarientFlux2Small — Method

InvarientFluxSmall(sSmol,mass12,mass22)

returns the value of the squared invarient flux with smalled 's' mandelstram variable (sSmol = s - (m1+m2)^2)

BoltzmannCollisionIntegral.InvarientFluxSmall — Method

InvarientFluxSmall(sSmol,mu12,mu22)

returns the value of the invarient flux with smalled 's' mandelstram variable (sSmol = s - (m1+m2)^2)

BoltzmannCollisionIntegral.Momentum3Value! — Method

Momentum3Value!(p3v,p3pv,p1v,p2v)

Takes set of random initial particle states 'p1v' and 'p2v' and random output states angles 'p3v[2:3]' and modifies outputs 'p3v' and 'p3pv' values with calculated output momentum and corrects angles if momentum is negative. Function also returns a two bools 'p3physical' and 'p3pphysical' indicating if p3 and p3p are physical momentum states given the inputs. Function also returns a Int 'NumStates' indicating the number of valid output states found.

Requrires normalised masses (mu1,mu2,mu3,mu4) to be defined in advance in Init.jl as const.

Examples

julia> mu1 = 1836.1528f0
julia> mu2 = 1836.1528f0
julia> mu3 = 1836.1528f0
julia> mu4 = 1836.1528f0
julia> p1v = [1f0, 0.5f0, 1.8f0]
julia> p2v = [2f0, 0.2f0, 0.7f0]
julia> p3v = [0f0, 0.3f0, 0.7f0]
julia> p3pv = zeros(Float32,3)
julia> p3pv .= p3v
julia> Momentum3Value!(p3v,p3pv,p1v,p2v,mu1,mu2,mu3,mu4)
(true,true,2)
julia> p3v
 3-element Vector{Float32}:
 2.04505
 0.3
 0.7
julia> p3pv
 3-element Vector{Float32}
 0.691423
 -0.3
 1.7

BoltzmannCollisionIntegral.PhaseSpaceFactors1! — Method

PhaseSpaceFactors1!(SMatrix,TMatrix,t3val,p1val,t1val,p2val,t2val,name1,name2)

Applies phase space volume element factors for 'SMatrix' and 'TMatrix' terms in order to correctly apply 'STSymmetry' corrections.

BoltzmannCollisionIntegral.PhaseSpaceFactors2! — Method

PhaseSpaceFactors2!(SMatrix,TMatrix,t3val,p1val,t1val,p2val,t2val)

To follow 'PhaseSpaceFactors1' and 'STSymmetry'. Corrects phase space factors on 'SMatrix' and 'TMatrix' for use in kinetic codes. Assumes f(x,p,μ)= constant

BoltzmannCollisionIntegral.RPointLogMomentum! — Method

RPointLogMomentum!(pu,pl,pv,num)

Edits the first element of pv with a random real-space momentum value between $10^{pl}$ and $10^{pu}$. This sample is chosen by first randomly picking a momentum bin in the range 1:num and then uniformly sampling a momentum point in real-space (rather than log10 space) between pl and pu which are the momentum values at start and end of that bin. Sampling is done such there will be a constant number of points per momentum-space volume. As the momentum space between $10^{pl}$ and $10^{pu}$ it is a spherical shell hence the correct sampling is $p = (U*(10^{pu})^3+(1-U)*(10^{pl})^3)^{1/3}$ with uniform $U ∈ [0~~1]$.

Assumes $f(x,p,μ)=f(x,\vec{p})*(2πp^2)=const$ in bin, therefore momentum space volume element is $\mathrm{d}p$ and as such uniform sampling corresponds to $U*10^{u}+(1-U)*10^{l}$ where $U$ is a uniform random number between 0 and 1.

If instead $f(x,\vec{p})=const$ in bin, momentum space volume element is $p^2 \mathrm{d}p$ and uniform sampling corresponds to $(10^pu)*\sqrt[3]{U+(1-U)*10^{3pl-3pu}}$ where $U$ is a uniform random number between 0 and 1.

BoltzmannCollisionIntegral.RPointSphereCosThetaPhi! — Method

RPointSphereThetaPhi!()

Assigns the second (cos(theta)) and third (phi) elements of 'a' with a randomly, uniformly sampled values of spherical angles cos(theta) and phi (phi normalised by pi).

BoltzmannCollisionIntegral.STMonteCarloAxi_MultiThread! — Method

STMonteCarloAxi_MultiThread!(SAtotal,TAtotal,SAtally,TAtally,p3v,p3pv,p1v,p2v,p3Max,t3MinMax,sigma,dsigmadt,Parameters,numTiterPerThread,numSiterPerThread)

Arguments

SAtotal::Array{Float32,6} : Array of stored integration totals for S matrix

TAtotal::Array{Float32,4} : Array of stored integration totals for T matrix

SAtally::Array{UInt32,5} : Array of stored integration tallies for S matrix

TAtally::Array{UInt32,4} : Array of stored integration tallies for T matrix

p3v::Vector{Float32} : Vector of momentum values for species 3

p3pv::Vector{Float32} : Second Vector of momentum values for species 3 for when two states are possible

p1v::Vector{Float32} : Vector of momentum values for species 1

p2v::Vector{Float32} : Vector of momentum values for species 2

p3Max::Array{Float32,5} : Array of maximum momentum values for species 3

t3MinMax::Array{Float32,6} : Array of minimum and maximum theta values for species 3

sigma::Function : Cross section function for the interaction

dsigmadt::Function : Differential cross section function for the interaction

Parameters::Tuple{Float32,Float32,Float32,Float32,Float32,Float32,Int64,Float32,Float32,Int64,Float32,Float32,Int64,Int64,Int64,Int64} : Tuple of parameters for the interaction

numTiterPerThread::Int64 : Number of T iterations per thread

numSiterPerThread::Int64 : Number of S iterations per thread

Output:

Argument arrays SAtotal,TAtotal,SAtally,TAtally are mutated to include the results of the Monte Carlo Integration.

Hidden Inputs (defined in Init.jl)

Domain Boundaries (defined as CONST) - p bounds and divisions for species 1,3,4 - theta divisions for species 1,3,4 ( bounds not needed as assumed [-1,1] ) - phi divisions for species 1,3,4 ( bounds not needed as assumed [0,2] )

Particle Masses (defined as CONST)

numTiterPerThread and numSiterPerThread as the number of T and S integrations to perform.

Calculation In Breif

Set up worker threads

Random Sample points in each of these domains

RandomPointSphere for theta and phi (for species 1,2,3)

RandomPointMomentum for p ( species 1,2 only )

Take random points (t3,h1,p1,p2,t1,t2,h3,h4) and calculate valid p3 point/points

Find position in local S and T arrays and allocated tallies and totals accordingly.

Update global S and T arrays with locks to prevent data races

BoltzmannCollisionIntegral.STMonteCarloAxi_Serial! — Method

STMonteCarloAxi_Serial!(SAtotal,TAtotal,SAtally,TAtally,p3v,p3pv,p1v,p2v,p3Max,t3MinMax,sigma,dsigmadt,Parameters,numTiter,numSiter)

Arguments

SAtotal::Array{Float32,6} : Array of stored integration totals for S matrix

TAtotal::Array{Float32,4} : Array of stored integration totals for T matrix

SAtally::Array{UInt32,5} : Array of stored integration tallies for S matrix

TAtally::Array{UInt32,4} : Array of stored integration tallies for T matrix

p3v::Vector{Float32} : Vector of momentum values for species 3

p3pv::Vector{Float32} : Second Vector of momentum values for species 3 for when two states are possible

p1v::Vector{Float32} : Vector of momentum values for species 1

p2v::Vector{Float32} : Vector of momentum values for species 2

p3Max::Array{Float32,5} : Array of maximum momentum values for species 3

t3MinMax::Array{Float32,6} : Array of minimum and maximum theta values for species 3

sigma::Function : Cross section function for the interaction

dsigmadt::Function : Differential cross section function for the interaction

Parameters::Tuple{Float32,Float32,Float32,Float32,Float32,Float32,Int64,Float32,Float32,Int64,Float32,Float32,Int64,Int64,Int64,Int64} : Tuple of parameters for the interaction

numTiter::Int64 : Number of T iterations

numSiter::Int64 : Number of S iterations

Output:

Argument arrays SAtotal,TAtotal,SAtally,TAtally are mutated to include the results of the Monte Carlo Integration.

Hidden Inputs (defined in Init.jl)

Domain Boundaries (defined as CONST) - p bounds and divisions for species 1,3,4 - theta divisions for species 1,3,4 ( bounds not needed as assumed [-1,1] ) - phi divisions for species 1,3,4 ( bounds not needed as assumed [0,2] )

Particle Masses (defined as CONST)

numTiter and numSiter as the number of T and S integrations to perform.

Calculation In Breif

Random Sample points in each of these domains

RandomPointSphere for theta and phi (for species 1,2,3)

RandomPointMomentum for p ( species 1,2 only )

Take random points (t3,h1,p1,p2,t1,t2,h3,h4) and calculate valid p3 point/points

Find position in S and T arrays and allocated tallies and totals accordingly.

BoltzmannCollisionIntegral.STSymmetry! — Method

STSymmetry!(SMatrix,TMatrix,t3val,p1val,t1val,p2val,t2val,mu1,mu2)

To follow 'PhaseSpaceFactors1'. Physical nature of binary interaction has certain symmetries. 'STSymmetry' uses these symmetries to improve MC sampling of 'SMatrix' and 'TMatrix'.

BoltzmannCollisionIntegral.SValue — Method

SValue(p3v,p1v,p2v,dsigmadt,mu1,mu2,mu3)

Returns Sval from MC integration based on initial momentum states p1v and p2v and final state p3v and differential cross section dsigmadt based on particle selection. Assumes f(x,p,μ)=constant over bin

BoltzmannCollisionIntegral.SpectraEvaluateMultiThread — Method

SpectraEvaluateMultiThread(userInputMultiThread)

Function to run the Monte Carlo integration of the S and T arrays in a multi-threaded environment. The function will run the Monte Carlo integration in parallel across the number of threads specified in the global variable nThreads. The function will then calculate the S and T matricies and save the results to a file.

BoltzmannCollisionIntegral.SpectraEvaluateSerial — Method

SpectraEvaluateSerial(userInputSerial)

Function to run the Monte Carlo integration of the S and T arrays in a serial environment. The function will run the Monte Carlo integration in serial and then calculate the S and T matricies and save the results to a file.

BoltzmannCollisionIntegral.TValue — Method

TValue(p1v,p2v,sigma,mu1,mu2)

returns Tval with its Tval from MC integration based on initial momentum states p1v and p2v and cross section sigma based on particle selection. If initial state fails sCheck, i.e. cannot generate a physical output state, Tval is set to 0f0. Assumes f(x,p,μ)=constant over bin

BoltzmannCollisionIntegral.deltaEVector — Method

deltaEVector(pr,mu)

Inputs a num+1 long Vector{Float32} of p grid boundries and the particle mu value (normalised mass) and returns a num long Vector{Float32} of average energy values per grid cell.

Examples

julia> deltaEVector([1.0f0, 10.0f0, 100.0f0, 1000.0f0], 1.0f0)
3-element Vector{Float32}:
 50.600693
 4951.15
 495001.16

BoltzmannCollisionIntegral.deltaVector — Method

deltaVector(valr)

Inputs a num+1 long Vector{Float32} quantitiy values (domain bounds) and returns a num long Vector{Float32} of differeces (domain widths).

Examples

julia> deltaVector([1.0f0, 10.0f0, 100.0f0, 1000.0f0])
3-element Vector{Float32}:
 9.0
 90.0
 900.0

BoltzmannCollisionIntegral.dsigmadt_ElePosPhoPho — Method

dsigmadt_ElePosPhoPho(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for electron positron annihilation to two photons. Berestetskii 1982 (88.4). Masses and momenta are normalised by the rest mass of the electron $m_{\text{Ele}}$ and the cross section is normalised by $3σ_T$.

\[\frac{dσ}{dt} = -\frac{1}{s(s-4)}\left(\left(\frac{1}{t-1}+\frac{1}{u-1}\right)^2+\left(\frac{1}{t-1}+\frac{1}{u-1}\right)-\frac{1}{4}\left(\frac{t-1}{u-1}+\frac{u-1}{t-1}\right)\right)\]

Arguments

sSmol::Float32 : $s - sBig$

sBig::Float32 : $(m_1+m_2)^2 = 4 ∴ s = sSmol + 4$

tSmol::Float32 : $t - tBig$

tBig::Float32 : $(m_3-m_1)^2 = 1 ∴ t = tSmol + 1$

uSmol::Float32 : $u - uBig$

uBig::Float32 : $(m2-m3)^2 = 1 ∴ u = uSmol + 1$

BoltzmannCollisionIntegral.dsigmadt_PhoPhoElePos — Method

dsigmadt_PhoPhoElePos(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for photon-photon annihilation to electron-positron pair. (Inverse proceess of electron positron annihilation to two photons). Masses and momenta are normalised by the rest mass of the electron $m_{\text{Ele}}$ and the cross section is normalised by $3σ_T$.

\frac{dσ}{dt} = -\frac{1}{s^2}\left(\left(\frac{1}{t-1}+\frac{1}{u-1}\right)^2+\left(\frac{1}{t-1}+\frac{1}{u-1}\right)-\frac{1}{4}\left(\frac{t-1}{u-1}+\frac{u-1}{t-1}\right)\right)

Arguments

sSmol::Float32 : $s - sBig$

sBig::Float32 : $(m_1+m_2)^2 = 0 ∴ s = sSmol$

tSmol::Float32 : $t - tBig$

tBig::Float32 : $(m_3-m_1)^2 = 1 ∴ t = tSmol + 1$

uSmol::Float32 : $u - uBig$

uBig::Float32 : $(m_2-m_3)^2 = 1 ∴ u = uSmol + 1$

BoltzmannCollisionIntegral.dsigmadt_SphSphSphSph — Method

dsigmadt_SphSphSphSph(sSmol,sBig,tSmol,tBig,uSmol,uBig)

returns the differential cross section for the binary interaction of hard spheres with normalised masses $m_1,m_2,m_3,m_4=m_{\text{Sph}}$. Normalised by $πR_{Sph}^2$.

\[\frac{dσ}{dt} = \frac{1}{s-4m_{\text{Sph}}^2}\]

Arguments

sSmol::Float32 : $s - sBig$

sBig::Float32 : $(m_1+m_2)^2=4m_{\text{Sph}}^2$

tSmol::Float32 : $t - tBig$

tBig::Float32 : $(m_3-m_1)^2=0$

uSmol::Float32 : $u - uBig$

uBig::Float32 : $(m_2-m_3)^2=0$

BoltzmannCollisionIntegral.fload_All — Method

fload_All(fileLocation,fileName)

Loads all the data stored in fileName stored at fileLocation.

Example

    (Run_Parameters, Stot,Ttot,Stal,Ttal,SMatrix,TMatrix,p3Max,t3MinMax,SConv,TConv) = fload_All(fileLocation,fileName);

Returns a tuple of the data stored in the file. The fields are as follows:

Run_Parameters : A tuple of the parameters used in the evaluation.

Stot : A 6D matrix totalling all the emission spectrum values sampled.

Ttot : A 4D matrix totalling all the absorption spectrum values sampled.

Stal : A 5D matrix of tallies of the number of emission spectrum values sampled.

Ttal : A 4D matrix of tallies of the number of absorption spectrum values sampled.

SMatrix : A 6D matrix of the emission spectrum.

TMatrix : A 4D matrix of the absorption spectrum.

p3Max : The maximum value of the momentum space variable p3 sampled for each bin. (Useful for correcting numerical diffusion)

t3MinMax : The minimum and maximum values of the momentum space variable t3 sampled for each bin. (Useful for correcting numerical diffusion)

SConv : A 6D matrix of the convergence of the emission spectrum compaired to the previous run with given Run_Parameters.

TConv : A 4D matrix of the convergence of the absorption spectrum compaired to the previous run with given Run_Parameters.

BoltzmannCollisionIntegral.location — Method

location(u,l,num,val)

Returns the index of the bin in which 'val' is contatined based on the 'num' of bins and their 'u' upper and 'l' lower bound.

Examples

julia> location(10f0,0f0,9,2f0)
2

BoltzmannCollisionIntegral.location_p3 — Method

location_p3(u,l,num,val)

Returns the index of the bin in which 'val' is contatined based on the 'num' of bins and their 'u' upper and 'l' lower bound including overflow and underflow possibilities. Overflow are assigned to num+1 while underflow are assigned to lowest bin i.e. 1.

Examples

julia> location_p3(10f0,1f0,9,2f0)
2
julia> location_p3(10f0,1f0,9,11f0) # overflow
10
julia> location_p3(10f0,1f0,9,0.5f0) # underflow
1

BoltzmannCollisionIntegral.location_t — Method

location_t(numt,val)

Returns the index of the bin in which the costheta 'val' is contatined based on the 'numt' of bins. Bounds [tl tu] are defined as CONST in Init.jl

Examples

julia> location_t(8,0.5f0)
6

BoltzmannCollisionIntegral.meanVector — Method

meanVector(valr)

Inputs a num+1 long Vector{Float32} of domain bounds and returns a num long Vector{Float32} of mean value in domain range.

Examples

julia> meanVector([1.0f0, 10.0f0, 100.0f0, 1000.0f0])
3-element Vector{Float32}:
 5.5
 55.0
 550.0

BoltzmannCollisionIntegral.prange — Method

prange(pl,pu,nump)

Returns a nump+1 long Vector{Float32} of p-space grid bounds NOT in Log10 space.

Examples

julia> prange(-5f0,4f0,9)
10-element Vector{Float32}:
 1.0e-5
 1.0e-4
 1.0e-3
 0.01
 0.1
 1.0
 10.0
 100.0
 1000.0
 10000.0

BoltzmannCollisionIntegral.sCheck — Method

sCheck(sSmol,sBig,mu3,mu4)

Returns 'true' if 's' mandelstram generated from inital system state can generate a physical output state.

BoltzmannCollisionIntegral.sigma_ElePosPhoPho — Method

sigma_ElePosPhoPho(sSmol,sBig)

returns the total cross section for electron positron annihilation to two photons. Berestetskii 1982 (88.6). Masses and momenta are normalised by the rest mass of the electron $m_{\text{Ele}}$ and the cross section is normalised by $3σ_T$.

\[σ = \frac{1}{4s^2(s-4)}\left((s^2+4s-8)\log\left(\frac{\sqrt{s}+\sqrt{s-4}}{\sqrt{s}-\sqrt{s-4}}\right)-(s+4)\sqrt{s(s-4)}\right)\]

Arguments

sSmol::Float32 : $s - sBig$

sBig::Float32 : $(m_1+m_2)^2 = 4 ∴ s = sSmol + 4$

BoltzmannCollisionIntegral.sigma_PhoPhoElePos — Method

sigma_PhoPhoElePos(sSmol,sBig)

returns the total cross section for photon-photon annihilation to electron-positron pair. Masses and momenta are normalised by the rest mass of the electron $m_{\text{Ele}}$ and the cross section is normalised by $3σ_T$.

\[σ = \frac{1}{2s^3}\left((s^2+4s-8)\log\left(\frac{\sqrt(s)+\sqrt(s-4)}{\sqrt(s)-\sqrt(s-4)}\right)-(s+4)\sqrt{s(s-4)}\right)\]

Arguments

sSmol::Float32 : $s - sBig$

sBig::Float32 : $(m_1+m_2)^2 = 0 ∴ s = sSmol$

BoltzmannCollisionIntegral.sigma_SphSphSphSph — Method

sigma_SphSphSphSph(sSmol,sBig)

returns the total cross section for the binary interaction of hard spheres with normalised masses (wrt electron mass) $m_1,m_2,m_3,m_4=m_\text{Sph}$. Normalised by $πR_{Sph}^2$.

\[σ = \frac{1}{2}\]

Arguments

sSmol::Float32 : s - sBig

sBig::Float32 : (m1+m2)^2

BoltzmannCollisionIntegral.trange — Method

trange(numt)

Returns a numt+1 long Vector{Float32} of theta-space grid bounds in terms of cos(theta). Upper and lower bounds [tl tu] are defined as CONST in Init.jl as [-1f0 1f0].

Examples

julia> trange(8)
9-element Vector{Float32}:
 -1.0
 -0.75
 -0.5
 -0.25
  0.0
  0.25
  0.5
  0.75
  1.0

BoltzmannCollisionIntegral.vectorLocation — Method

vectorLocation(pu,pl,nump,numt,vector)

Returns a tuple of bin location for (log10momentum,cos(theta)) based on an input 'vector' and bounds 'u,l' of their domains and the 'num' of uniformly spaced bins. costheta bounds [tl tu] are defined as CONST in Init.jl

Examples

```julia-repl julia> vectorLocation(4f0,-5f0,9,8,[1f0,0.5f0,1.5f0]) (5,6)